# openapi/kayan_openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Kayan Orphan Care — AI Agent Platform (Mock)",
    "description": "Prototype backend simulating جمعية كيان للايتام beneficiary management. Multi-channel (Voice/SIP + WhatsApp), Arabic-first. Each route is an agent tool.",
    "version": "1.0.0"
  },
  "paths": {
...
    "/enrollments": {
      "post": {
        "tags": [
          "7 · الاعتماد والصرف | Enrollment & Disbursement"
        ],
        "summary": "Enrol an approved beneficiary into a program",
        "description": "Creates the enrollment for an ACCEPTED request and generates the monthly disbursement schedule. Returns 409 if the request was not accepted or is already enrolled.",
        "operationId": "create_enrollment_enrollments_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/EnrollIn"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/beneficiary/{beneficiary_id}/enrollments": {
      "get": {
        "tags": [
          "7 · الاعتماد والصرف | Enrollment & Disbursement"
        ],
        "summary": "List a beneficiary's program enrollments",
        "description": "Every program the beneficiary is enrolled in, with amounts and status.",
        "operationId": "list_enrollments_beneficiary__beneficiary_id__enrollments_get",
        "parameters": [
          {
            "name": "beneficiary_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Beneficiary Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/beneficiary/{beneficiary_id}/disbursements": {
      "get": {
        "tags": [
          "7 · الاعتماد والصرف | Enrollment & Disbursement"
        ],
        "summary": "List scheduled and paid disbursements",
        "description": "The monthly transfer schedule with totals paid and upcoming — this answers 'كم اتحول لي وكم باقي'.",
        "operationId": "list_disbursements_beneficiary__beneficiary_id__disbursements_get",
        "parameters": [
          {
            "name": "beneficiary_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Beneficiary Id"
            }
          },
          {
            "name": "status",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "examples": [
                "scheduled"
              ],
              "title": "Status"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/disbursements/{disbursement_id}/approve": {
      "post": {
        "tags": [
          "7 · الاعتماد والصرف | Enrollment & Disbursement"
        ],
        "summary": "Approve a disbursement for payment (التعميد بالصرف)",
        "description": "Moves a scheduled disbursement to approved so finance can pay it.",
        "operationId": "approve_disbursement_disbursements__disbursement_id__approve_post",
        "parameters": [
          {
            "name": "disbursement_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Disbursement Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ApproveDisbIn"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/disbursements/{disbursement_id}/pay": {
      "post": {
        "tags": [
          "7 · الاعتماد والصرف | Enrollment & Disbursement"
        ],
        "summary": "Execute payment to the beneficiary's IBAN",
        "description": "Records a bank transfer against the disbursement and sends an SMS notice. Blocked (409) if the beneficiary has no IBAN on file.",
        "operationId": "pay_disbursement_disbursements__disbursement_id__pay_post",
        "parameters": [
          {
            "name": "disbursement_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Disbursement Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/beneficiary/{beneficiary_id}/payments": {
      "get": {
        "tags": [
          "7 · الاعتماد والصرف | Enrollment & Disbursement"
        ],
        "summary": "Payment history",
        "description": "All settled payments with bank reference numbers.",
        "operationId": "list_payments_beneficiary__beneficiary_id__payments_get",
        "parameters": [
          {
            "name": "beneficiary_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Beneficiary Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/finance/disbursement-run": {
      "get": {
        "tags": [
          "7 · الاعتماد والصرف | Enrollment & Disbursement"
        ],
        "summary": "Upcoming disbursement run",
        "description": "Everything due in the next N days across all beneficiaries — the monthly payment run the finance team executes.",
        "operationId": "disbursement_run_finance_disbursement_run_get",
        "parameters": [
          {
            "name": "days",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 120,
              "minimum": 1,
              "default": 30,
              "title": "Days"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/sponsorships": {
      "get": {
        "tags": [
          "8 · الكفالات والفعاليات | Sponsorships & Events"
        ],
        "summary": "List sponsorships (الكفالات)",
        "description": "Sponsor-to-beneficiary kafala records, restricted or unrestricted.",
        "operationId": "list_sponsorships_sponsorships_get",
        "parameters": [
          {
            "name": "beneficiary_id",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Beneficiary Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/events": {
      "get": {
        "tags": [
          "8 · الكفالات والفعاليات | Sponsorships & Events"
        ],
        "summary": "List events and activities (الفعاليات والانشطة)",
        "description": "Association events with capacity and registration counts.",
        "operationId": "list_events_events_get",
        "parameters": [
          {
            "name": "status",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "examples": [
                "scheduled"
              ],
              "title": "Status"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/events/{event_id}/register": {
      "post": {
        "tags": [
          "8 · الكفالات والفعاليات | Sponsorships & Events"
        ],
        "summary": "Register a beneficiary for an event",
        "description": "Registers an approved beneficiary for an event; blocked when the event is full or already completed.",
        "operationId": "register_event_events__event_id__register_post",
        "parameters": [
          {
            "name": "event_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Event Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RegisterEventIn"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/beneficiary/{beneficiary_id}/history": {
      "get": {
        "tags": [
          "9 · السجل الشامل | Beneficiary 360"
        ],
        "summary": "Complete beneficiary history (السجل الشامل)",
        "description": "THE ONE CALL that returns everything about a beneficiary in one place: file status and completeness, household, documents, finances, every support request with its decision, enrollments, disbursements, payments, sponsorships, tickets and channel sessions. Use it to open any conversation with full context.",
        "operationId": "history_beneficiary__beneficiary_id__history_get",
        "parameters": [
          {
            "name": "beneficiary_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Beneficiary Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "EnrollIn": {
        "properties": {
          "support_request_id": {
            "type": "string",
            "title": "Support Request Id",
            "examples": [
              "SR-20001"
            ]
          },
          "type": {
            "type": "string",
            "title": "Type",
            "default": "one_time",
            "examples": [
              "one_time",
              "monthly_recurring"
            ]
          },
          "months": {
            "type": "integer",
            "maximum": 36.0,
            "minimum": 1.0,
            "title": "Months",
            "description": "Number of monthly instalments for recurring support",
            "default": 1,
            "examples": [
              12
            ]
          },
          "start_date": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Start Date",
            "examples": [
              "2026-08-01"
            ]
          }
        },
        "type": "object",
        "required": [
          "support_request_id"
        ],
        "title": "EnrollIn"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
    }
  }
}